Move BlogList to presenational components. Add BlogPage container.
